refactor: [Menu] - 불필요한 String 타입의 변수 제거 [CommandLineApplication] -패키지 구조 변경
