+++ src/main/java/com/programmers/SpringVoucherServiceApplication.java
@@ -1,6 +1,5 @@
 package com.programmers;
 
-import com.programmers.voucher.CommandLineApplication;
 import org.slf4j.Logger;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
